requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
db-dtypes==1.1.1
et-xmlfile==1.1.0
Flask==2.2.3
Flask-Cors==3.0.10
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.18.1
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.10.0
google-cloud-bigquery-storage==2.20.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.128
llama-index==0.5.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.6
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
pandas-gbq==0.19.2
proto-plus==1.22.2
protobuf==4.23.1
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
pydata-google-auth==1.8.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
tenacity==8.2.2
tiktoken==0.4.0
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.3
yarl==1.9.2
youtube-transcript-api==0.6.0
